More namespace API updates.
Added PrefixMapping[Impl] to record the prefix mappings of a Model.
Models now extend PrefixMapping, so any model can now do a setNsPrefix.
Those prefixes are passed to the Writer before it writes.

collection.rdf almost certainly has a bug in its use. The daml namespace
was wrong but this isn't the complete problem.

RDFWriter's have been forced to have an enquiry method for finding out
if there's already a prefix for some URI. At the moment only the XML
writers implement it.


git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-CVS/jena2/trunk@1109632 13f79535-47bb-0310-9956-ffa450edef68