updates to allow OntModels to delegate reification to their distinguished base element.
I made sure MultiUnion passed the Graph tests - this exposed some problems,
ie it didn't, because a listener on an OntModel got poked twice for each
update. To fix this, MultiUnions suppress notifyAdd and notifyDelete,
because the MultiUnion shares the base graph's event handler, and *that*
will produce notifications.

It turned out also that simply delegating the MultiUnion reifier to the
base didn't work. Tests failed. It turns out that ModelComs [and hence
OntModelImpls] grab the underlying Graph's reifier when they are constructed,
but a MultiUnion's reifier can change as the base changes, and starts out
as a default non-base reifier; the MultiUnion base was not set when the
MultiUnion was created.

To fix this, the OntModelImpl constructor now passes the base model down
to the code that constructs its underlying MultiUnion graph, and that
code puts the base graph into the union, so that it's available when
the OntModelImpl's parent class constructor [ie ModelCom's] runs and
grabs the reifier.

BaseInfGraph and its descendants delegate their reifiers to their raw
graphs.


git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-CVS/jena2/trunk@1111352 13f79535-47bb-0310-9956-ffa450edef68