Fixed problem in GraphExtract where extracting into an existing graph which already has some triple(s) (R, p, o) for [sub]root R would not add any triples for R from the graph being extracted from.
This was because the test for "already seen" was the existence in the
destination graph of (R, ANY, ANY), which works fine if that graph
starts out empty - but because we expose and use the "update" form,
isn't safe.

Resorted to using an "active node" set. Also refactored the core code
into a nested class to allow the elimination of most of the parameter
passing and result returning.


git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-CVS/jena2/trunk@1111718 13f79535-47bb-0310-9956-ffa450edef68