JENA-135
Added a JenaWriterRdfJson, it extends JenaWriterBase however RiotWriter.writeRDFJSON(...) method wants an OutputStream.
Searching what's the best way to go around this.

A Jena writer has methods to write on a Writer as well as an OutputStream. Writers are discouraged. Would a JenaWriterRdfJson which works with only OutputStream be acceptable?

git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Jena2/ARQ/trunk@1187316 13f79535-47bb-0310-9956-ffa450edef68