JENA-36 and JENA-155
This is to make sure we do not have log messages (either at WARN or ERROR level) when we run the test suite in Eclipse or via mvn test command line.

There are two log4j properties file: one is used by the test suite and it's located in src/test/resources/log4j.properties. The other one is src/main/java/jena-log4j.properties and it is included in the .jar file (because we want command lines to have logging enabled). I usually have just one log4j.properties in src/test/resources and do not ship jars with log4j in it. If that is needed, a better location for jena-log4j.properties is src/main/resources. I am not going do to this now.

I agree on the principle of not having spurious errors or warning when we execute the test suite (and this is the reason why I am committing this).
However, I did not find this task trivial as it should have been. We could benefit from a common and shared approach towards logging, error messages, System.out|err, in the code and in the tests. It there is one, I am not aware of it and I am not able to infer it from the sources. At the moment IMHO there is a too high degree of diversity (which some could call caos). :-)

I am not proposing to do anything about this now, let's focus on the release. However a discussion on what would be the right thing to do and sharing a best practice is possible.

git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Jena2/jena/trunk@1197540 13f79535-47bb-0310-9956-ffa450edef68