Major performance boost to SPARQL TSV parser (25x performance improvement seen on a 3.5 million result file - 6s down from 150s)
Change to achieve this is to make the parser avoid calling into NodeFactory.parseNode() where possible by sniffing terms to see if they look like URIs/Blank Nodes based on the first character or so and assuming they are and just generating a Node directly in those cases.  We now only call into NodeFactory.parseNode() for literals because literals are complicated and we want all the nice escape handling the parser will give us.

git-svn-id: https://svn.apache.org/repos/asf/jena/trunk/jena-arq@1356519 13f79535-47bb-0310-9956-ffa450edef68