Provide a registry for query and update serializers
This commit introduces a registry for query and update serializers that allows
users to register serializers for custom syntaxes which allows extended syntaxes
not understood by ARQ to be printed.  With the current API the serialization of
queries and updates is hard coded to use the ARQ serialization.  When no appropriate
serializer is registered the code falls back to the old hard coded behaviour.




git-svn-id: https://svn.apache.org/repos/asf/jena/trunk@1625498 13f79535-47bb-0310-9956-ffa450edef68