Add various additional Leviathan Functions (JENA-507)
- pythagoras() calculates the length of the hypotenuse of a right
  angled triangle given the lengths of the other two sides
- degrees-to-radians() and radians-to-degrees()
- cos() and cos-1() i.e. acos

As part of this commit some improvements were made to Loader to deal
with class name strings that contain illegal characters to the Java
language specification.  So for example degrees-to-radians would be an
invalid Java class name so it gets escaped to degressToRadians and the
loader looks for that class instead.