Work on getting rat:check to pass
Added license headers to various files in jena-core that needed them.
Added excludes for test resources because historically Jena has large
numbers of small test files and it is unclear if we can just blanket add
the license headers to them.  They are all covered by the original SGA
to the ASF or by subsequent ICLAs so from an IP perspective are clean
anyway