Work towards dynamic loading of readers and writers in Hadoop RDF IO
This commit adds the first stage of work towards supporting entirely
dynamic loading of readers and writers in the Hadoop RDF IO module.
This is intended for use with the TriplesReader, QuadsReader and
TriplesOrQuadsReader to allow them to dynamically select the appropriate
reader to use for a given input file.  Making the loading dynamic allows
users to plug in custom RDF formats to ARQ and register them with the
Hadoop RDF IO module such that no core code changes are required to
start supporting a new input format under Hadoop.