Add support for dynamically selected triples/quads output
This commit adds QuadsOutputFormat, TriplesOutputFormat and
TriplesOrQuadsOutputFormat which are counterparts to the existing
equivalent input formats.  These output formats dynamically select the
actual record writer to use based upon the target output filename.  This
is useful for jobs where the output filename may be user supplied so the
desired output format is not known in advance.