Explictly add <extensions>true</extensions> to OSGI POM
On some build servers an issue has been reported with the bundle
packaging not being recognized.  In principal this should be picked up
from the bundle plugin which should inherit the extensions setting from
the parent but perhaps is not.  This is thus an attempt to see if adding
it explicitly works around this issue.