JENA-901 use a LinkedList instead of ArrayList
for activeInterpreters. They could be closed in arbitrary order,
and also this avoids a large ArrayList remaining after a big burst
of concurrent queries.

Simplify test. Check activeInterpreters is empty - this didn't
happen as I had forgotten to call it.hasNext() (which any use of
an iterator would normally do).

Test now with MAX=1024 and 128M find() calls.