Fix a bug in OpAsQuery.asQuery() (JENA-954)
Some modifiers that can wrap a project (distinct, reduced and slice)
were being incorrectly moved from the sub-query to the parent query when
trying to round trip a SPARQL query through algebra form.

This commit modifies OpAsQuery so that the conversion of those algebra
operators checks whether we're already inside a project (and thus the
occurrence of these operators means we have a sub-query) and if so
converts the operator and its sub-tree into a sub-query on the parent
query.

With this commit test cases based upon JENA-954 all now pass