Don't merge extends when looking for inlining oppurtunities (JENA-780)
In checking the new optimization it was noticed that this introduces a
regression in the query JENA-779 that originally spawned this proposal.
This was because we used OpExtend.extend() rather than OpExtend.create()
which could result in other optimizations being blocked.

We now use OpExtend.create() instead and add a test to check that we
don't merge extends.