Don't inline assignments in some unsafe cases (JENA-780)
There are some cases where inlining is unsafe around moving an
assignment from within an n-ary expression e.g. join, leftjoin, union,
minus etc because the assignments value could be dependent on the branch
of the query it occurs in and moving it outside the branching operator
could change the semantics of its evaluation.

This commit disables inlining in those cases and adds extra test cases
to ensure that those cases are not subject to inlining.

Conflicts:
	jena-arq/src/main/java/org/apache/jena/sparql/algebra/optimize/TransformEliminateAssignments.java