New Dataset with operation record that allows abort from transaction
Introduces a package in jena-arq named o.a.j.sparql.core.journaling which includes the DatasetGraph implementation DatasetGraphWithRecord, which wraps another DatasetGraph. The wrapped DatasetGraph must have MRSW locking semantics. DatasetGraphWithRecord uses types in o.a.j.s.c.j to keep a record of operations performed against it, in the form of ReversibleOperationRecord. This record can be used to abort an open transaction against a dataset wrapped by DatasetGraphWithRecord, or for other purposes.