Add new rdfdiff tool to command line tools
This commit adds a new rdfdiff tool to the command line tools, this has
the same arguments as rdfcompare except that it actually produces
information about the differences rather than simply indicating if the
two files are equal.

It is placed in the ARQ command line tools because coping with blank
nodes re-uses the Closure class from ARQ.

It has not been heavily tested against blank node containing graphs but
has been verified with plenty of concrete graphs.