Fix for JENA-1063
This commit changes the behaviour of QueryEngineHTTP to avoid an
undesireable behaviour from Apache HTTP Client which we use for our HTTP
communications.

HTTP Client tries to re-use connections by default
which means that it must finish consuming responses before it can close
the InputStream associated with a specific HTTP response.  However this
can cause a hang and leave both the client and remote server stuck doing
unecessary work since most of the time we are going to clean up the HTTP
Client anyway so leaving the connections available for re-use makes no
sense for us.

The change is essentially to check whether we are going to clean up the
HTTP Client and if so do that first so that when we clean up the
associated InputStream the underlying connection is already closed and
it can't and won't waste time trying to consume the remaining response.