Implemented fn:round (xpath3) and fn:round-to-half-even
I implemented fn:round-to-half-even. While implementing it, I realized that the fn:round function was not following the xpath3 directive and I thus changed it. Still as the round function (without fn) is used in SPARQL and SPARQL 1.1 seems to be based on Xpath2 (true?), I created a new roundxpath3 function that is implementing the xpath3 specs.