Make runtime regex pattern errors a ExprEvalException.
ExprEvalException does not have a stacktrace

A regex error a "syntax error" is a parse failure if it occurs
for a fix regex pattern and a logged runtime eval failure if compiled
dynamically during query execution.  The FILTER will eval to false.

No stacktrace is printed for the logged runtime exception.