Migrated blank node hashing from MD5 to SHA-256
MD5 is a deprecated hashing algorithm, and even though it is not used on sensitive data in the context of Jena, its usage is picked up by security softwares as a security flaw. This may reduce the incentive to use Jena in commercial products, and computing SHA-256 hashes is not prohibitively more expensive than MD5.