[JENA-1911] New UI
This commit replaces the old BackboneJS by Vue.JS, using the Vue CLI utility
and npm to build this new Maven module. It also uses a Maven build plug-in to
invoke NPM.

License and Notice files were updated, as well as the build settings to generate
these files. Other changes include modifications to the jena-fuseki-webapp
module on how it loads the UI files, and other minor changes required to have
the new UI (e.g. ignore YARN/NPM files in RAT).

See JENA-1911 and the two pull requests created for this change for more.