Improves TransformPathFlattenAlgebra
Adds ability to flatten some further p{N,M} variants e.g. p{N,} that
were not previously handled.  Ensures that the N step is done first and
always to/from the bounded end of the path (if any).

Lots of additional test cases around this to figure out correct
behaviours and find bugs in the implementation.