Optimise SELECT * WHERE { GRAPH ?g { } } for TDB2 (GH-1639)
Apply the usage of the new distinct by key iterator to
SolverLibTDB.graphNames() path ensuring that the rest of the logical
flow there continues to work as before. Added explanatory comments about
the choices and optimisations involved.

Moved repeated logic for selecting a suitable index actually into the
TupleTable class and simplified some code as a result