GH-2150 Fix for bad query AST builder generated
In some cases using the query builder API to create a query/update could
result in an apparently valid query (when printing as a string) but the
actual AST was bad.  When attempting to actually execute the
query/update locally ARQ would fail to compile it as a result.

This commit fixes BuildElementVisitor to avoid an oversimplification of
some AST elements, specifically an ElementGroup containing a single
ElementFilter that was leading to the bad AST.