HttpAction transaction fixes (GH-2543)
In looking into whether the prefixes service PR #2538 could be
refactored to reduce code duplication I noticed that if an Action is
registered directly as servicing a URL, rather than being associated
with a Dataset, then if the Action tries to use any of the transaction
handling of HttpAction it can result in NPEs because some access to the
transactional context was not checking for non-null transactional
objects.  This commit addresses that problem and should allow #2538 to
be refactored.