GH-2738: Performance improvements to CacheSimple Slight performance improvements to org.apache.jena.atlas.lib.cache.CacheSimple: - optimized bit-operations by using a cache size, that is always a power of 2 - removed already BiConsumer<K, V> dropHandler), which is used nowhere - added more tests for CacheSimple - fixed a small bug, where putting a key with the same hash as an existing one does not overridde the existing entry, when the values are equal. - added new module jena-benchmarks-shadedJena51 to be able to perform regression tests againt the previous release - added JHM test org.apache.jena.atlas.lib.cache.TestCaches in jena-benchmarks-jmh - added more detailed javadoc comments to Cache interface
Downside: The fixed cache size must always be a power of 2. If the given size is already a power of two it will be used as fixed size for the cache, otherwise the next larger power of two will be used. (e. g. minimumSize = 10 results in 16 as fixed size for the cache)