GH-2753: Improve customising Fuseki args
Improves how Fuseki arguments are customised, in particular makes the
ServerConfig class public adding the ability to store extra
configuration on it.

It also changes the previously private static applyServerArgs() method
into a protected instance method so users can extend FusekiMain and
override it to use that to customise the FusekiServer.Builder further.

Unit tests are added that validate that custom arguments can now be
passed through CLI inputs and used to influence the Server being built.